requirements.txt
langchain
pydantic
python-dotenv
langchain-community
langchain-core
PyGithub
openai
langchain_openai